+++ src/ArrayProvider.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * @see       https://github.com/zendframework/zend-config-aggregator for the canonical source repository
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2015-2016 Mateusz Tymek (http://mateusztymek.pl)
+ * @license   https://github.com/zendframework/zend-config-aggregator/blob/master/LICENSE.md New BSD License
+ */
+
+namespace Zend\ConfigAggregator;
+
+/**
+ * Provider that returns the array seeded to itself.
+ *
+ * Primary use case is configuration cache-related settings.
+ */
+class ArrayProvider
+{
+    /**
+     * @var array
+     */
+    private $config;
+
+    /**
+     * @param array $config
+     */
+    public function __construct(array $config)
+    {
+        $this->config = $config;
+    }
+
+    /**
+     * @return array
+     */
+    public function __invoke()
+    {
+        return $this->config;
+    }
+}