+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 0.2.0 - 2017-01-11
+
+### Added
+
+- [#2](https://github.com/zendframework/zend-config-aggregator/pull/2) adds a
+  new `ArrayProvider`, which accepts an array to its constructor, and returns
+  it when invoked. This can be used to provide in-line array configuration when
+  feeding the `ConfigAggregator` instance.
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
 ## 0.1.0 - 2016-12-08
 
 Initial release.